+++ src/app/background/mod.rs
@@ -7,7 +7,7 @@ pub mod bootstrap;
 pub mod load_library_task;
 pub mod library_scan_task;
 
-pub mod person_thumbnail;
+pub mod person_thumbnail_task;
 
 pub mod photo_clean_task;
 pub mod photo_detect_faces_task;
