+++ src/app/background/person_thumbnail_task.rs
